# Make content_tag a bit faster