This speeds up `content_tag` by about 17%

1. Replacing two `Set` lookups with a single `Hash` lookup
2. Avoiding a `String#dup` for values which don't include `"`

One possible downside is that this requires freezing `TagHelper::BOOLEAN_ATTRIBUTES` and `TagHelper::TAG_PREFIXES`. I'm not sure if we expected users to be able to modify these.

## Benchmark

https://gist.github.com/jhawthorn/0d3b20e0156885799960da62a5aac23a

**Before**

```
$ be ruby benchmarks/content_tag.rb
Warming up --------------------------------------
              simple    24.325k i/100ms
          with quote    23.599k i/100ms
     octicons_helper     7.433k i/100ms
Calculating -------------------------------------
              simple    294.251k (± 5.2%) i/s -      1.484M in   5.056557s
          with quote    283.826k (± 3.3%) i/s -      1.440M in   5.077635s
     octicons_helper     84.693k (± 4.7%) i/s -    423.681k in   5.014594s

```

**After**

```
$ be ruby benchmarks/content_tag.rb
Warming up --------------------------------------
              simple    33.045k i/100ms
          with quote    31.069k i/100ms
     octicons_helper    10.563k i/100ms
Calculating -------------------------------------
              simple    366.623k (± 3.1%) i/s -      1.851M in   5.052713s
          with quote    327.993k (± 8.0%) i/s -      1.647M in   5.060007s
     octicons_helper    113.388k (± 5.2%) i/s -    570.402k in   5.045369s
```