@kaspth 2019-10-12
👍 I'm fine with this. Though I have been thinking we should deprecate `content_tag` in favor of `tag.div` because it can handle both `tag :div` and `content_tag :div` in one method 😊