@jhawthorn 2019-10-25
This should make all of `tag.div`, `tag :div` and `content_tag :div` faster 🙂